requirements.txt
asgiref==3.3.1
astroid==2.4.2
atomicwrites==1.4.0
attrs==20.3.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
coverage==5.5
cryptography==3.3.1
defusedxml==0.7.0rc2
Django==3.1.5
django-cors-headers==3.7.0
django-extensions==3.1.1
django-templated-mail==1.1.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
djoser==2.1.0
drf-yasg==1.20.0
factory-boy==3.2.0
Faker==6.6.2
freezegun==1.1.0
idna==2.10
inflection==0.5.1
iniconfig==1.1.1
isort==5.7.0
itypes==1.2.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
oauthlib==3.1.0
packaging==20.9
pluggy==0.13.1
py==1.10.0
pycparser==2.20
PyJWT==2.0.1
pylint==2.6.0
pyparsing==2.4.7
pytest==6.2.3
pytest-django==4.2.0
pytest-factoryboy==2.1.0
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.5
requests==2.25.1
requests-oauthlib==1.3.0
ruamel.yaml==0.17.0
ruamel.yaml.clib==0.2.2
selenium==3.141.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.0.3
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.3
wrapt==1.12.1
